fix layout/file system
