+++ src/app/(admin)/admin/login/page.tsx
@@ -4,14 +4,50 @@ import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { createClient } from '@/utils/supabase/client';
+import { useEffect } from 'react';
 
-export default function Login() {
+export default function AdminLogin() {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
   const [error, setError] = useState<string | null>(null);
   const [loading, setLoading] = useState(false);
+  const [checkingAuth, setCheckingAuth] = useState(true);
   const router = useRouter();
   const supabase = createClient();
+
+  // Auth guard - verificar si hay una sesión activa
+  useEffect(() => {
+    const checkSession = async () => {
+      try {
+        const { data: { session } } = await supabase.auth.getSession();
+        
+        if (session) {
+          // Si hay una sesión activa, verificar si es administrador
+          const res = await fetch('/api/admin/validate', {
+            method: 'GET',
+            headers: {
+              'Content-Type': 'application/json',
+            },
+          });
+          
+          if (res.ok) {
+            const validationData = await res.json();
+            // Redirigir al dashboard si es un admin válido
+            router.push(`/admin/dashboard/${validationData.institution_id}`);
+            return;
+          }
+          // Si no es admin, cerrar sesión
+          await supabase.auth.signOut();
+        }
+      } catch (error) {
+        console.error('Error checking session:', error);
+      } finally {
+        setCheckingAuth(false);
+      }
+    };
+    
+    checkSession();
+  }, [router, supabase.auth]);
 
   const handleLogin = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -19,55 +55,69 @@ export default function Login() {
     setLoading(true);
 
     try {
-      // First, call your custom authentication endpoint
-      const res = await fetch('/api/auth/login', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ email, password }),
+      // Autenticar directamente con Supabase
+      const { error: authError } = await supabase.auth.signInWithPassword({
+        email,
+        password,
       });
 
-      const data = await res.json();
+      if (authError) {
+        setError('Credenciales inválidas');
+        return;
+      }
+
+      // Verificar si el email pertenece a una institución
+      const res = await fetch('/api/admin/validate', {
+        method: 'GET',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+      });
+
+      const validationData = await res.json();
       
       if (!res.ok) {
-        setError(data.error || 'Error al iniciar sesión');
+        // Si hay un error, cerrar la sesión que acabamos de iniciar
+        await supabase.auth.signOut();
+        setError(validationData.error || 'No tienes permisos de administrador');
         return;
       }
       
-      // If the custom endpoint authentication was successful, update Supabase's auth state
-      // This ensures the navbar and other components that rely on Supabase auth state are updated
-      if (data.session) {
-        // Set the session in Supabase client
-        await supabase.auth.setSession({
-          access_token: data.session.access_token,
-          refresh_token: data.session.refresh_token
-        });
-        
-        // Navigate to home page
-        router.push('/home');
-      } else {
-        // If no session was returned, show an error
-        setError('Error al iniciar sesión: No se recibió una sesión válida');
-      }
+      // Si la validación fue exitosa, redirigir al dashboard de administración
+      router.push(`/admin/dashboard/${validationData.institution_id}`);
     } catch (error: unknown) {
-      const errorMessage = error instanceof Error ? error.message : 'An error occurred during login';
+      const errorMessage = error instanceof Error ? error.message : 'Ocurrió un error durante el inicio de sesión';
       setError(errorMessage);
     } finally {
       setLoading(false);
     }
   };
 
+  if (checkingAuth) {
+    return (
+      <main className="flex items-center justify-center bg-white text-gray-800">
+        <div className="flex items-center justify-center">
+          <svg className="animate-spin h-8 w-8 text-gray-500" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+            <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+            <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+          </svg>
+        </div>
+      </main>
+    );
+  }
+
   return (
-    <main className="min-h-screen flex items-center justify-center bg-white text-gray-800">
+    <main className="flex items-center justify-center text-gray-800">
       <div className="w-full max-w-sm p-6 border border-gray-200 rounded-xl shadow-sm">
         <h1 className="text-xl font-semibold text-center mb-1">
-          Iniciar sesión en <span className="font-bold text-gray-900">InterviewLab</span>
+          Panel de Administración <span className="font-bold text-gray-900">InterviewLab</span>
         </h1>
-        <p className="text-sm text-center text-gray-500 mb-6">Ingresa tus credenciales</p>
+        <p className="text-sm text-center text-gray-500 mb-6">Acceso exclusivo para administradores</p>
         <form className="mt-8 space-y-6" onSubmit={handleLogin}>
           <div className="space-y-4 rounded-md">
             <div>
               <label htmlFor="email" className="block text-sm font-medium">
-                Email address
+                Email
               </label>
               <input
                 id="email"
@@ -82,7 +132,7 @@ export default function Login() {
             </div>
             <div>
               <label htmlFor="password" className="block text-sm font-medium">
-                Password
+                Contraseña
               </label>
               <input
                 id="password"
@@ -133,9 +183,9 @@ export default function Login() {
           </button>
         </form>
         <p className="text-xs text-center text-gray-500 mt-6">
-          ¿No tienes cuenta?{' '}
-          <Link href="/register" className="text-gray-800 underline">
-            Regístrate
+          ¿No eres administrador?{' '}
+          <Link href="/login" className="text-gray-800 underline">
+            Acceso de usuario
           </Link>
         </p>
       </div>
