+++ src/app/(auth)/home/page.tsx
@@ -10,7 +10,7 @@ export default async function HomePage() {
   }
 
   return (
-    <main className="min-h-screen flex flex-col items-center justify-start bg-white text-gray-800 px-4 py-16">
+    <main className="flex flex-col items-center justify-center text-gray-800 px-4">
       <div className="text-center space-y-4 mb-12 mt-12">
         <h1 className="text-3xl font-bold">Hola, {data.user.email}</h1>
         <p className="text-gray-600">Bienvenido a InterviewLab</p>
@@ -50,22 +50,15 @@ export default async function HomePage() {
             <div className="flex items-start">
               <div className="bg-black text-white rounded-full w-8 h-8 flex items-center justify-center flex-shrink-0 mt-1">3</div>
               <div className="ml-4">
-                <h3 className="font-semibold text-lg">Realiza la entrevista</h3>
-                <p className="text-gray-700">Participa en una entrevista interactiva con nuestro entrevistador virtual Alejandro.</p>
-              </div>
-            </div>
-            <div className="flex items-start">
-              <div className="bg-black text-white rounded-full w-8 h-8 flex items-center justify-center flex-shrink-0 mt-1">4</div>
-              <div className="ml-4">
-                <h3 className="font-semibold text-lg">Mejora tus habilidades</h3>
-                <p className="text-gray-700">Practica tantas veces como necesites para ganar confianza y mejorar tus respuestas.</p>
+                <h3 className="font-semibold text-lg">Inicia la entrevista</h3>
+                <p className="text-gray-700">Comienza la simulación de entrevista y responde a las preguntas como lo harías en una entrevista real.</p>
               </div>
             </div>
           </div>
         </section>
 
         <div className="text-center">
-          <Link href="/entrevista" className="inline-block px-8 py-4 bg-black text-white rounded-md hover:bg-gray-800 transition-all duration-300 transform hover:translate-y-[-2px] font-medium">
+          <Link href="/entrevista" className="inline-block bg-gray-900 text-white px-6 py-3 rounded-md font-medium hover:bg-black transition">
             Comenzar una entrevista
           </Link>
         </div>
