+++ src/app/(public)/layout.tsx
@@ -1,0 +1,18 @@
+import Navbar from "@/components/navbar";
+import Footer from "@/components/Footer";
+
+export default function PublicLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <Navbar />
+      <div className="flex-grow overflow-y-auto content-center justify-center py-4 bg-radial from-white from-40% to-gray-100">
+        {children}
+      </div>
+      <Footer />
+    </>
+  );
+}
