+++ src/app/(public)/terms/page.tsx
@@ -2,7 +2,7 @@
 
 export default function TermsPage() {
   return (
-    <main className="min-h-screen py-12 px-6 bg-white">
+    <main className="px-6 bg-white">
       <div className="max-w-3xl mx-auto mt-12">
         <h1 className="text-3xl font-bold mb-8 text-gray-900">Términos y Condiciones</h1>
