--- src/app/admin/layout.tsx
@@ -1,34 +1,0 @@
-'use client';
-
-import { Geist, Geist_Mono } from "next/font/google";
-import "../globals.css";
-import AdminNavbar from "@/components/AdminNavbar";
-import Footer from "@/components/Footer";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export default function AdminLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="es">
-      <body className={`${geistSans.variable} ${geistMono.variable} font-sans antialiased flex flex-col min-h-screen`}>
-        <AdminNavbar />
-        <div className="flex-grow">
-          {children}
-        </div>
-        <Footer />
-      </body>
-    </html>
-  );
-}
