+++ src/components/ResumeForm.tsx
@@ -51,7 +51,7 @@ export function ResumeForm({ onComplete }: ResumeFormProps) {
   };
 
   return (
-    <div className="bg-white p-8 rounded-lg shadow-lg mb-8 w-full border border-gray-100 transition-all duration-300">
+    <div className="bg-white p-8 rounded-lg shadow-sm mb-8 w-full border border-gray-100 transition-all duration-300">
       <h2 className="text-2xl font-bold mb-6 text-black border-b pb-3">Subir Archivos</h2>
       <form onSubmit={handleSubmit} className="space-y-6">
         <div className="mb-5">
